MicroserviciosDocker/ControlUsuarios/requirements.txt

Click==7.0
enum34==1.1.6
Flask==0.12.2
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.10
SQLAlchemy==1.1.11
Werkzeug==0.14.1
PyJWT==1.7.1
requests

